refactor: Introduce useStaking
