+++ src/composables/index.ts
@@ -2,6 +2,7 @@ import useConnectableRadix from './useConnectableRadix'
 import useHomeModal from './useHomeModal'
 import useNativeToken from './useNativeToken'
 import useRadix from './useRadix'
+import useStaking from './useStaking'
 import useSidebar from './useSidebar'
 import useTokenBalances from './useTokenBalances'
 import useWallet from './useWallet'
@@ -11,6 +12,7 @@ export {
   useHomeModal,
   useNativeToken,
   useRadix,
+  useStaking,
   useSidebar,
   useTokenBalances,
   useWallet
